pruntime: Add safe mode

+++ crates/phactory/src/prpc_service.rs
@@ -226,7 +226,13 @@ impl<Platform: pal::Platform + Serialize + DeserializeOwned> Phactory<Platform> 
         let counters = self.runtime_state()?.storage_synchronizer.counters();
         blocks.retain(|b| b.block_header.number >= counters.next_block_number);
 
-        let mut last_block = counters.next_block_number - 1;
+        let last_block = blocks
+            .last()
+            .map(|b| b.block_header.number)
+            .unwrap_or(counters.next_block_number - 1);
+
+        let safe_mode = self.args.safe_mode;
+
         for block in blocks.into_iter() {
             info!("Dispatching block: {}", block.block_header.number);
             let state = self.runtime_state()?;
@@ -234,17 +240,19 @@ impl<Platform: pal::Platform + Serialize + DeserializeOwned> Phactory<Platform> 
                 .storage_synchronizer
                 .feed_block(&block, state.chain_storage.inner_mut())
                 .map_err(from_display)?;
+            if safe_mode {
+                continue;
+            }
             info!("State synced");
             state.purge_mq();
             self.check_requirements();
             self.handle_inbound_messages(block.block_header.number)?;
-            last_block = block.block_header.number;
 
-            if let Err(e) = self.maybe_take_checkpoint(last_block) {
+            // NOTE: Please make sure don't save checkpoints in safe mode if you are refactoring codes.
+            if let Err(e) = self.maybe_take_checkpoint(block.block_header.number) {
                 error!("Failed to take checkpoint: {:?}", e);
             }
         }
-
         Ok(pb::SyncedTo {
             synced_to: last_block,
         })
